prototype/requirements.txt
# Package requirements for generated kedro project
Python <3.11, >=3.7 
kedro==0.18.8
kedro-mlflow==0.11.8
kedro-viz==6.1.0
